refactor(browser): extract browser Rsbuild assembly
